Updating how routing works
This changes all the server endpoints to his /api/v1/ instead of their normal endpoints. This was done
to avoid conflicts on page reloading where you might hit json routes instead of the actual page data.

This also makes some adjustments to the angular routing. The server will redirect all non '/api' links to the home page.
When it does this is stores the url that the user is trying to hit and passes that back to the the front end via cookie.

The front end then retrives the cookie and tries to redirect the user in Angular to the route that was requested. Or
redirects them to the homepage is it's not a real route.