Adding ability to editing users
This adds the ability to edit users for admin enabled users.

This tweaks a number of different areas on both the server side user controller and on the client side.

On the server side it restricts the get all users endpoint only to admins. On the client side, it shows the link to
get users only for admin users and not for anyone else. Additionally, it will return an error is the user hitting
the /api/v1/users endpoint is not an admin user. The view is single page based allow you to save and update mutiple users
withou the page reloading.