Adding the create a flow bit to add concepts
This commit moves the ability to create a flow diagram to the add concept page. It requires the users to be signed
up or logged in to create a flow diagram. The user can now create a flow diagram and add it to their concept from the
same page.

There's a number of changes to the add concept controller to handle the ability to add an image
when you add a new concept. Those should be reviewed.

Other things included in this commit are just removing the references to the /draw route that were in there previously
for testing purposes.